+++ larpmanager/locale/en/LC_MESSAGES/django.po
@@ -4510,6 +4510,24 @@ msgstr ""
 msgid "Short url"
 msgstr ""
 
+msgid "Archive mail"
+msgstr ""
+
+msgid "This page shows all emails, both queued and sent."
+msgstr ""
+
+msgid "Recipient"
+msgstr ""
+
+msgid "Subject"
+msgstr ""
+
+msgid "Body"
+msgstr ""
+
+msgid "Sent"
+msgstr ""
+
 msgid ""
 "This page shows membership badges. You can edit them, assign them, create "
 "new ones."
@@ -7372,6 +7390,9 @@ msgstr ""
 msgid "Membership check"
 msgstr ""
 
+msgid "Archive"
+msgstr ""
+
 msgid "Prologues type"
 msgstr ""
 
@@ -7597,7 +7618,7 @@ msgstr ""
 msgid "Profile updated"
 msgstr ""
 
-msgid "Mail sent!"
+msgid "Mail added to queue!"
 msgstr ""
 
 msgid "Answer submitted!"
