+++ larpmanager/templates/trans.html
@@ -306,6 +306,7 @@
 {% blocktrans %}LarpManager{% endblocktrans %}
 {% blocktrans %}Profile{% endblocktrans %}
 {% blocktrans %}Template{% endblocktrans %}
+{% blocktrans %}Archive{% endblocktrans %}
 {% blocktrans %}Event{% endblocktrans %}
 {% blocktrans %}Run{% endblocktrans %}
 {% blocktrans %}Roles{% endblocktrans %}
@@ -376,3 +377,4 @@
 {% blocktrans %}Cache Reloading{% endblocktrans %}
 {% blocktrans %}Send e-mail{% endblocktrans %}
 {% blocktrans %}Form{% endblocktrans %}
+{% blocktrans %}Archive{% endblocktrans %}
