+++ larpmanager/views/exe/assoc.py
@@ -20,8 +20,6 @@
 
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.core.paginator import Paginator
-from django.db.models import Case, When, Value, IntegerField, OuterRef, Subquery
 from django.db.models import Prefetch
 from django.shortcuts import redirect, render
 from django.utils.translation import gettext_lazy as _
@@ -40,18 +38,11 @@ from larpmanager.forms.member import (
     ExeProfileForm,
 )
 from larpmanager.models.access import AssocRole
-from larpmanager.models.accounting import (
-    AccountingItem,
-    AccountingItemOther,
-    PaymentInvoice,
-    RefundRequest,
-)
 from larpmanager.models.association import Association, AssocText
 from larpmanager.models.base import FeatureModule, Feature
 from larpmanager.models.event import (
     Run,
 )
-from larpmanager.models.member import Membership
 from larpmanager.cache.role import check_assoc_permission
 from larpmanager.utils.common import (
     get_feature,
@@ -172,114 +163,6 @@ def exe_features_off(request, num):
     ctx = check_assoc_permission(request, "exe_features")
     exe_features_go(request, ctx, num, on=False)
     return redirect("exe_features")
-
-
-def exe_paginate(request, ctx, typ, selrel=None, show_runs=True, afield=None, subtype=None):
-    cls = typ
-    if hasattr(typ, "objects"):
-        cls = typ.objects
-    else:
-        typ = typ.model
-
-    elements = cls.filter(assoc_id=ctx["a_id"]).order_by("-created")
-    if 'hide' in [f.name for f in typ._meta.get_fields()]:
-        elements = elements.filter(hide=False)
-
-    run = -1
-    page = 1
-    search = ""
-    size = 20
-
-    if request.method == "POST":
-        run = int(request.POST.get("run", "-1"))
-        page = int(request.POST.get("page", 0))
-        search = request.POST.get("search", "")
-        size = int(request.POST.get("size", 20))
-
-    if run >= 0:
-        if run == 0:
-            if afield:
-                kwargs = {"{0}__{1}__isnull".format(afield, "run"): True}
-                elements = elements.filter(**kwargs)
-            else:
-                elements = elements.filter(run__isnull=True)
-        else:
-            if afield:
-                kwargs = {"{0}__{1}".format(afield, "run"): run}
-                elements = elements.filter(**kwargs)
-            else:
-                elements = elements.filter(run=run)
-
-    # ~ if run and run >= 0:
-    # ~ if run == 0:
-    # ~ elements = elements.filter(run__isnull=True)
-    # ~ else:
-    # ~ if afield:
-    # ~ kwargs = {'{0}__{1}'.format(afield, 'run'): run}
-    # ~ elements = elements.filter(**kwargs)
-    # ~ else:
-    # ~ elements = elements.filter(run = run)
-
-    if issubclass(typ, AccountingItem):
-        elements = elements.select_related("member")
-
-    if issubclass(typ, PaymentInvoice):
-        elements = elements.annotate(
-            is_submitted=Case(
-                When(status=PaymentInvoice.SUBMITTED, then=Value(0)),
-                default=Value(1),
-                output_field=IntegerField(),
-            )
-        )
-        elements = elements.order_by("is_submitted", "-created")
-
-    if issubclass(typ, RefundRequest):
-        elements = elements.prefetch_related("member__memberships")
-        elements = elements.order_by("-status", "-updated")
-
-        memberships = Membership.objects.filter(member_id=OuterRef("member_id"), assoc_id=ctx["a_id"]).order_by("id")[
-            :1
-        ]
-
-        # for c in ctx["list"]:
-        # for mb in c.member.memberships.all():
-        # if mb.assoc_id == ctx["a_id"]:
-        # c.membership = mb
-        elements = elements.annotate(credits=Subquery(memberships.values("credit")))
-
-    if subtype == "credits":
-        elements = elements.filter(oth=AccountingItemOther.CREDIT)
-    elif subtype == "tokens":
-        elements = elements.filter(oth=AccountingItemOther.TOKEN)
-
-    if selrel:
-        for e in selrel:
-            elements = elements.select_related(e)
-
-    if search:
-        elements = elements.filter(search__icontains=search)
-
-    paginator = Paginator(elements, per_page=size)
-    page = min(page, paginator.num_pages)
-    ctx["pagin"] = paginator
-    ctx["page"] = page
-    ctx["size"] = size
-    ctx["size_range"] = [20, 50, 100, 150, 200, 250, 500, 1000, 2000, 5000]
-
-    if page > 0:
-        ctx["list"] = paginator.page(page)
-    else:
-        ctx["list"] = elements.all()
-
-    ctx["show_runs"] = show_runs
-    ctx["runs"] = [(-1, _("All")), (0, "Assoc")] + [
-        (r.id, str(r))
-        for r in Run.objects.filter(event__assoc_id=ctx["a_id"], end__isnull=False)
-        .select_related("event")
-        .order_by("-end")
-    ]
-    ctx["run_sel"] = run
-    ctx["search"] = search
 
 
 @login_required
